Move compact JWS serialization to a submodule
